chore: separated models into seprate files to keep it organized
