+++ lib/pages/scout/Widget Factories/AthleteDetailsWidgetExports.dart
@@ -1,0 +1,2 @@
+export 'package:ax_dapp/pages/scout/Widget%20Factories/MLBAthleteDetailsWidget.dart';
+export 'package:ax_dapp/pages/scout/Widget%20Factories/NFLAthleteDetailsWidget.dart';
