+++ lib/pages/scout/models/SportsModel/AthleteScoutModelExports.dart
@@ -1,0 +1,2 @@
+export 'package:ax_dapp/pages/scout/models/SportsModel/MLBAthleteScoutModel.dart';
+export 'package:ax_dapp/pages/scout/models/SportsModel/NFLAthleteScoutModel.dart';
